Add option to exclude top-level tag plots (similar to old version)
It wouldn't have been the end of the world to keep them, but the plot titles
for the tags were often just the internal tag name (e.g. "lit") which looked
weird in the paper.  Also, the "lit" tag plot was shown three times since we
have separate plots for search/qa/tables, which was also weird.  So this commit
adds an option to exclude, and paper_plots.py uses that option.